Optimize the implementation by using binary search
Speeds up the sorting by performing an O(log n) search instead of O(n)
search for each element.

Also, this has an optimization to avoid doing the search when the
element is already at the correct place.

This optimization allowed increasing the number of elements in some
tests by a significant amount (from 1000 to 10_000_000) and the tests
still take roughly the same amount of time.